# vulkan: workaround for #10710 and #12147 16 bit unpack8 bug