@stduhpf 2025-03-20
Works like a charm, thanks @netrunnereve. It makes it even stranger that the bug still isn't fixed at the driver level, the logic between unpacking from 16 bits or 32 bits should be pretty much the same.