@netrunnereve 2025-03-20
> It makes it even stranger that the bug still isn't fixed at the driver level, the logic between unpacking from 16 bits or 32 bits should be pretty much the same.

The SPIRV is different as glslc doesn't optimize out the additional unpacks and the 32 bit one is treated as an unpack of 4 8 bit values into 4 floats. In the assembly I'm seeing that it correctly loads the bottom 8 bits into the .x slot and the next set of 8 bits in the .y slot for a 32 bit input, but if I use a 16 bit input it somehow loads the identical bottom bits into *both* the .x and .y slots. I mean they should have some sort of automated regression test for this... right?