@netrunnereve 2025-03-22
> I'm not convinced we'll ever be able to remove this again, but let's keep hoping.

Even if they fix it for RDNA GCN probably still has this bug and Windows GCN drivers are finished. So yeah I guess we'll have to avoid 16 bit unpack8s from now on.